Modularize onboarding flow into multi-step screens
