+++ app/index.tsx
@@ -1,3 +1,3 @@
-import OnboardingFlow from './onboarding';
+import OnboardingFlow from '../screens/onboarding';
 
 export default OnboardingFlow;
