+++ screens/onboarding/styles.ts
@@ -1,0 +1,112 @@
+import { StyleSheet, Dimensions } from 'react-native';
+
+export const { width: SCREEN_W, height: SCREEN_H } = Dimensions.get('window');
+
+export const styles = StyleSheet.create({
+  screen: { flex: 1 },
+  page: {
+    width: SCREEN_W,
+    height: SCREEN_H,
+    paddingHorizontal: 16,
+    paddingTop: 80,
+    alignItems: 'center',
+  },
+  card: {
+    width: '100%',
+    maxWidth: 520,
+    backgroundColor: 'rgba(255,255,255,0.06)',
+    borderRadius: 18,
+    padding: 18,
+    borderWidth: StyleSheet.hairlineWidth,
+    borderColor: 'rgba(255,255,255,0.15)',
+  },
+  backBtn: {
+    position: 'absolute',
+    top: 16,
+    left: 16,
+    zIndex: 10,
+    padding: 8,
+    borderRadius: 999,
+    backgroundColor: 'rgba(255,255,255,0.08)',
+    borderWidth: StyleSheet.hairlineWidth,
+    borderColor: 'rgba(255,255,255,0.18)',
+  },
+  backPressed: { transform: [{ scale: 0.98 }] },
+  skipBtn: {
+    position: 'absolute',
+    top: 16,
+    right: 16,
+    zIndex: 10,
+    paddingVertical: 8,
+    paddingHorizontal: 12,
+    borderRadius: 999,
+    backgroundColor: 'rgba(255,255,255,0.08)',
+    borderWidth: StyleSheet.hairlineWidth,
+    borderColor: 'rgba(255,255,255,0.18)',
+  },
+  skipPressed: { transform: [{ scale: 0.98 }] },
+  skipText: { color: '#e5e7eb', fontWeight: '600' },
+  lottie: {
+    width: 160,
+    height: 160,
+    alignSelf: 'center',
+    marginBottom: 8,
+  },
+  title: { color: 'white', fontSize: 28, fontWeight: '700', letterSpacing: 0.2 },
+  subtitle: {
+    color: '#cbd5e1',
+    fontSize: 16,
+    marginTop: 6,
+    marginBottom: 14,
+  },
+  bullets: { gap: 8, marginTop: 6 },
+  bullet: { color: '#e5e7eb', fontSize: 16, lineHeight: 22 },
+  dotsRow: {
+    position: 'absolute',
+    bottom: 112,
+    left: 0,
+    right: 0,
+    flexDirection: 'row',
+    justifyContent: 'center',
+    gap: 8,
+  },
+  dot: {
+    width: 8,
+    height: 8,
+    borderRadius: 8,
+    backgroundColor: 'rgba(255,255,255,0.3)',
+  },
+  dotActive: { backgroundColor: '#60a5fa' },
+  cta: {
+    position: 'absolute',
+    bottom: 32,
+    left: 16,
+    right: 16,
+    paddingVertical: 14,
+    borderRadius: 14,
+    alignItems: 'center',
+    backgroundColor: '#60a5fa',
+    borderWidth: StyleSheet.hairlineWidth,
+    borderColor: 'rgba(255,255,255,0.18)',
+  },
+  ctaPressed: { transform: [{ scale: 0.98 }] },
+  ctaText: { color: 'white', fontWeight: '700', fontSize: 16 },
+  fineprint: {
+    color: '#9ca3af',
+    fontSize: 12,
+    marginTop: 14,
+  },
+  chartWrap: { marginTop: 6, alignItems: 'center' },
+  chartLabel: { color: '#e5e7eb', marginBottom: 8, fontSize: 12 },
+  chartLabelsRow: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    marginTop: 6,
+  },
+  chartLabelItem: { flex: 1, alignItems: 'center' },
+  chartXLabel: { color: '#cbd5e1', fontSize: 12 },
+  chartXLabelActive: { color: 'white', fontWeight: '700' },
+  chartValue: { color: '#e5e7eb', marginTop: 8, fontSize: 12 },
+  chartFootnote: { color: '#9ca3af', marginTop: 2, fontSize: 10 },
+});
+
